requirements.txt
absl-py==0.15.0
aiohttp==3.8.5
aiosignal==1.2.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
apturl==0.5.2
asn1crypto==0.24.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
beautifulsoup4==4.12.2
braceexpand==0.1.7
Brlapi==0.6.6
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
Cython==3.0.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
easydict==1.9
ez-setup==0.9
filelock==3.4.1
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
httplib2==0.9.2
huggingface-hub==0.4.0
hydra-core==1.1.0
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
kornia==0.5.0
lama @ file:///home/tracking/Inpainting/dependencies/dist/lama-0.1.0-py3-none-any.whl
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.36.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.7
MarkupSafe==1.0
matplotlib==3.3.4
multidict==5.2.0
netifaces==0.10.4
networkx==2.5.1
nibabel==3.2.2
numba==0.53.1
numpy==1.19.5
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
omegaconf==2.1.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pexpect==4.2.1
Pillow==8.4.0
platformdirs==2.4.0
progress==1.6
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.1.0
pyRFC3339==1.0
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-debian==0.1.32
pytorch-lightning==1.2.9
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==6.0.1
regex==2023.8.8
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.9
sacremoses==0.0.53
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Shapely==1.8.5.post1
simplejson==3.13.2
six==1.15.0
soupsieve==2.3.2.post1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.10
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.6.12
tokenizers==0.12.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchmetrics==0.2.0
torchvision==0.9.0+cu111
tqdm==4.64.1
transformers==4.18.0
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.3
virtualenv==20.14.1
wadllib==1.3.2
webdataset==0.2.48
Werkzeug==2.0.3
wldhx.yadisk-direct==0.0.6
wrapt==1.12.1
xkit==0.0.0
yarl==1.7.2
zipp==3.6.0
zope.interface==4.3.2
